Confirmation for user deletion fixed.

+++ plantproject/views.py
@@ -208,17 +208,15 @@ def comment_delete(request, project_pk, card_pk, comment_pk):
 
 def delete_project(request, project_pk):
     project = get_object_or_404(PlantProject, pk=project_pk)
-    project.delete()
-    return HttpResponseRedirect(reverse('home'))
-
-    # if request.user == project.creator:
-    #     if request.method == 'POST':
-    #         project.delete()
-    #         return redirect('userprofile:user-profile', pk=request.user.pk)
-    #     else:
-    #         return redirect('project-view', pk=project_pk)
-    # else:
-    #     return render(request, 'error_page.html', {'message': 'You are not authorized to delete this project.'})
+    
+    if request.user == project.creator:
+        if request.method == 'POST':
+            project.delete()
+            return redirect('userprofile:user-profile', pk=request.user.pk)
+        else:
+            return redirect('project-view', pk=project_pk)
+    else:
+        return render(request, 'error_page.html', {'message': 'You are not authorized to delete this project.'})
 
 
 class EditProjectView(UpdateView):
